fix: change recommanded wallet to connect scenes
