+++ packages/connector/src/scenes/wallets.ts
@@ -2,32 +2,59 @@ import { ccc } from "@ckb-ccc/ccc";
 import { html } from "lit";
 import { repeat } from "lit/directives/repeat.js";
 import { WalletWithSigners } from "../types";
+import { recommendedWallets, WalletInfo } from "./recommandedWallet";
 
 export function generateWalletsScene(
   wallets: WalletWithSigners[],
   onWalletSelected: (wallet: ccc.Wallet) => unknown,
   onSignerSelected: (wallet: ccc.Wallet, signer: ccc.SignerInfo) => unknown,
 ) {
-  return [
-    "Connect Wallet",
-    repeat(
-      wallets,
-      (wallet) => wallet.name,
-      (wallet) => html`
+
+  const missingWallets = recommendedWallets.filter(
+    (recommendedWallet) =>
+      !wallets.some((wallet) => wallet.name === recommendedWallet.name),
+  );
+
+  const additionalButtons = missingWallets.map(
+    (wallet: WalletInfo) => html`
+      <div>
+        <img class="wallet-icon" src=${wallet.icon} alt=${wallet.name} />
+        <span class="mb-2">${wallet.name}</span>
         <button
           class="btn-primary mb-1"
-          @click=${async () => {
-            if (wallet.signers.length === 1) {
-              onSignerSelected(wallet, wallet.signers[0]);
-            } else {
-              onWalletSelected(wallet);
-            }
+          @click=${() => {
+            window.open(wallet.downloadLink, "_blank");
           }}
         >
-          <img src=${wallet.icon} alt=${wallet.name} />
-          ${wallet.name}
+          Download ${wallet.name}
         </button>
-      `,
-    ),
+      </div>
+    `,
+  );
+  
+  return [
+    "Connect Wallet",
+    html`
+      ${repeat(
+        wallets,
+        (wallet) => wallet.name,
+        (wallet) => html`
+          <button
+            class="btn-primary mb-1"
+            @click=${async () => {
+              if (wallet.signers.length === 1) {
+                onSignerSelected(wallet, wallet.signers[0]);
+              } else {
+                onWalletSelected(wallet);
+              }
+            }}
+          >
+            <img src=${wallet.icon} alt=${wallet.name} />
+            ${wallet.name}
+          </button>
+        `,
+      )}
+      ${additionalButtons}
+    `,
   ];
 }
